all station display and envi temp signals

+++ senv/signals.py
@@ -42,7 +42,7 @@ class Signal(QObject):
     tkm_get_auth = pyqtSignal(list)
     tkm_get_train_auth = pyqtSignal(int)
     tkm_get_beacon = pyqtSignal(int,int)
-    tkm_get_envi_temp = pyqtSignal(int)
+    tkm_get_envi_temp = pyqtSignal(float)
     tkm_get_sales = pyqtSignal(int)
     tkm_get_pass_count = pyqtSignal(int)
     tkm_get_block = pyqtSignal(int)
